+++ src/ui/components/Card/components/ContentStackedCard.tsx
@@ -4,7 +4,7 @@ import BaseCard from './BaseCard';
 import useThemedStyles from '../../../hooks/useThemedStyles';
 import type { BaseCardProps, ContentStackedProps } from '../types';
 import Text from '../../../components/Text/Text';
-import type { ThemeContextType } from '../../../../src/theme/types';
+import type { ThemeContextType } from '../../../theme/types';
 
 interface TextContentProps {
   title?: string;
